pyproject.toml
[tool.poetry]
name = "invoice-parser"
version = "0.0.0"
description = "GCS powered tool for invoice parsing and information extraction"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10"

tqdm = "*"
pandas = "*"
pyarrow = "*"
google-cloud-documentai = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
